backend/requirements.txt
fastapi
uvicorn
google-generativeai
firebase-admin
stripe
python-multipart
pillow
python-dotenv
requests
